+++ changelog.md
@@ -8,6 +8,9 @@ Codion Change Log
 - Control.Builder.value() added.
 - AbstractControlBuilder added, related refactoring.
 - Control and ToggleControl.copyBuilder() added.
+### is.codion.swing.framework.ui
+- EntityEditPanel now overrides EntityEditComponentPanel.onException() and propagates the exception to the correct exception handler.
+- EntityTablePanel.onException() now propagates the exception to the correct exception handler.
 
 ## 0.17.21
 ### is.codion.swing.common.model